+++ clients/ctl/admin-ui/__tests__/features/taxonomy-transform.test.tsx
@@ -14,16 +14,19 @@ describe("data category transform", () => {
                 description: "Account's city level address data.",
                 label: "Account City",
                 value: "account.contact.city",
+                is_default: true,
               },
             ],
             description: "Contact data related to a system account.",
             label: "Account Contact Data",
             value: "account.contact",
+            is_default: true,
           },
         ],
         description: "Data related to a system account.",
         label: "Account Data",
         value: "account",
+        is_default: true,
       },
       {
         children: [
@@ -32,11 +35,13 @@ describe("data category transform", () => {
             description: "Data used to manage access to the system.",
             label: "Authentication Data",
             value: "system.authentication",
+            is_default: true,
           },
         ],
         description: "Data unique to, and under control of the system.",
         label: "System Data",
         value: "system",
+        is_default: true,
       },
       {
         children: [
@@ -46,6 +51,7 @@ describe("data category transform", () => {
               "Data derived from user provided data or as a result of user actions in the system.",
             label: "Derived Data",
             value: "user.derived",
+            is_default: true,
           },
           {
             children: [
@@ -55,6 +61,7 @@ describe("data category transform", () => {
                   "Data provided or created directly by a user that is not identifiable.",
                 label: "User Provided Non-Identifiable Data",
                 value: "user.provided.nonidentifiable",
+                is_default: true,
               },
               {
                 children: [
@@ -63,6 +70,7 @@ describe("data category transform", () => {
                     description: "Age range data.",
                     label: "User Provided Non-Specific Age",
                     value: "user.provided.identifiable.non_specific_age",
+                    is_default: true,
                   },
                   {
                     children: [],
@@ -70,24 +78,28 @@ describe("data category transform", () => {
                       "Data related to the individual's political opinions.",
                     label: "Political Opinion",
                     value: "user.provided.identifiable.political_opinion",
+                    is_default: true,
                   },
                 ],
                 description:
                   "Data provided or created directly by a user that is linked to or identifies a user.",
                 label: "User Provided Identifiable Data",
                 value: "user.provided.identifiable",
+                is_default: true,
               },
             ],
             description:
               "Data provided or created directly by a user of the system.",
             label: "User Provided Data",
             value: "user.provided",
+            is_default: true,
           },
         ],
         description:
           "Data related to the user of the system, either provided directly or derived based on their usage.",
         label: "User Data",
         value: "user",
+        is_default: true,
       },
     ]);
   });
@@ -100,6 +112,7 @@ describe("data category transform", () => {
         value: "anonymous_user",
         description:
           "An individual that is unidentifiable to the systems. Note - This should only be applied to truly anonymous users where there is no risk of re-identification",
+        is_default: true,
         children: [],
       },
       {
@@ -107,6 +120,7 @@ describe("data category transform", () => {
         value: "citizen_voter",
         description:
           "An individual registered to voter with a state or authority.",
+        is_default: true,
         children: [],
       },
       {
@@ -114,6 +128,7 @@ describe("data category transform", () => {
         value: "commuter",
         description:
           "An individual that is traveling or transiting in the context of location tracking.",
+        is_default: true,
         children: [],
       },
     ]);